[Modify] Add meta pkg and separate pkgs(bringup, toolbox) correctly
