+++ mikata_arm/CMakeLists.txt
@@ -1,0 +1,4 @@
+cmake_minimum_required(VERSION 2.8.3)
+project(mikata_arm)
+find_package(catkin REQUIRED)
+catkin_metapackage()
