+++ mikata_arm/package.xml
@@ -1,24 +1,21 @@
 <?xml version="1.0"?>
 <package>
-  <name>mikata_arm_bringup</name>
+  <name>mikata_arm</name>
   <version>0.9.0</version>
-  <description>The mikata_arm_bringup package</description>
+  <description>Sample programs and utilities for the Dynamixel MikataArm (meta package)</description>
   <license>Apache-2.0</license>
   <author email="jp.support@robotis.com">ROBOTIS Japan Technical Support</author>
   <maintainer email="jp.support@robotis.com">ROBOTIS Japan Technical Support</maintainer>
   <url type="website">https://github.com/ROBOTIS-JAPAN-GIT/dynamixel_mikata_arm</url>
   <url type="repository">https://github.com/ROBOTIS-JAPAN-GIT/dynamixel_mikata_arm</url>
   <buildtool_depend>catkin</buildtool_depend>
-  <build_depend>roscpp</build_depend>
-  <build_depend>mikata_arm_msgs</build_depend>
-  <build_depend>mikata_arm_toolbox</build_depend>
-  <build_depend>dynamic_reconfigure</build_depend>
-  <build_depend>dynamixel_sdk</build_depend>
-  <build_depend>eigen</build_depend>
-  <run_depend>roscpp</run_depend>
-  <run_depend>mikata_arm_msgs</run_depend>
-  <run_depend>dynamic_reconfigure</run_depend>
   <run_depend>dynamixel_sdk</run_depend>
-  <run_depend>eigen</run_depend>
-  
+  <run_depend>mikata_arm_bringup</run_depend>
+  <run_depend>mikata_arm_msgs</run_depend>
+  <run_depend>mikata_arm_toolbox</run_depend>
+
+  <export>
+    <metapackage/>
+  </export>
 </package>
+
