+++ mikata_arm_bringup/CMakeLists.txt
@@ -7,15 +7,16 @@ find_package(catkin REQUIRED COMPONENTS
   roscpp
   dynamixel_sdk
   mikata_arm_msgs
+  mikata_arm_toolbox
   dynamic_reconfigure
   cmake_modules
-  )
+)
 find_package(Eigen3 REQUIRED)
 
 generate_dynamic_reconfigure_options(
   cfg/Position.cfg
   cfg/Velocity.cfg
-  )
+)
 
 catkin_package(DEPENDS EIGEN3)
 
@@ -28,6 +29,7 @@ include_directories(
 add_executable(bringup src/bringup.cpp)
 add_executable(gui_control src/gui_control.cpp)
 
+add_dependencies(bringup ${catkin_EXPORTED_TARGETS})
 add_dependencies(gui_control ${PROJECT_NAME}_gencfg)
 
 target_link_libraries(bringup ${catkin_LIBRARIES})
