+++ mikata_arm_toolbox/CMakeLists.txt
@@ -7,16 +7,22 @@ find_package(catkin REQUIRED COMPONENTS
   roscpp
   dynamixel_sdk
   cmake_modules
-  )
+)
 find_package(Eigen3 REQUIRED)
 
-catkin_package(DEPENDS EIGEN3)
+catkin_package(
+  INCLUDE_DIRS include
+  LIBRARIES mikata_arm_toolbox
+  DEPENDS EIGEN3
+)
 
 include_directories(
   include
   ${catkin_INCLUDE_DIRS}
   ${EIGEN3_INCLUDE_DIRS}
 )
+
+add_library(mikata_arm_toolbox src/dxl_setup.cpp)
 
 add_executable(dxl_factory_reset src/dxl_factory_reset.cpp)
 add_executable(dxl_setup src/dxl_setup.cpp)
@@ -25,3 +31,4 @@ add_executable(cli_control src/cli_control.cpp)
 target_link_libraries(dxl_factory_reset ${catkin_LIBRARIES})
 target_link_libraries(dxl_setup ${catkin_LIBRARIES})
 target_link_libraries(cli_control ${catkin_LIBRARIES})
+
