--- mikata_arm_toolbox/include/motion_data.dat
@@ -1,6 +1,0 @@
-Initial_Pose ,5.0
-p1 ,4.0
-p4 ,2.0
-p5 ,1.0
-Rest_Pose ,6.0
-# LOOP
