--- mikata_arm_toolbox/include/teaching_data.dat
@@ -1,7 +1,0 @@
-Initial_Pose ,0.00153398 ,-0.00306796 ,0.00153398 ,-0.00153398 ,
-Rest_Pose ,0.00153398 ,-1.81163 ,1.40819 ,-1.72573 ,
-p1 ,0.0291456 ,0.0352816 ,1.27781 ,-1.34684 ,
-p2 ,-0.0322136 ,-0.716369 ,-0.081301 ,0.785398 ,
-p3 ,-0.859029 ,-0.0444854 ,0.843689 ,-1.01703 ,
-p4 , , , , ,0.5
-p5 , , , , ,0.0
